requirements.txt
streamlit==1.37.1
pandas==2.2.2
pillow==10.4.0
pyyaml==6.0.2
